feature: gaps in section-order
This adds another feature to section-order. It allows the user to add a
constant gap to the addresses.

Use case:
In a special case flashtool bootloader that is located in sram for jlink,
I found the following symbol ordering:
 * =0x2000000  !.vectors TEXT, RODATA, DATA, BSS
 * !__stack_start
 * +0x1000 !__stack_end
 * !__heap_start
 * =0x2040000 !__heap_end
 * !__flash_tool_cfg

So this naturally translates to:
  --section-order="=0x20000000 !.vectors TEXT RODATA DATA  BSS !__bss_end !__stack_start +0x1000 !__stack_end !__heap_start =0x20040000 !__heap_end !__flash_tool_cfg"

The address mangling for splitting the sram area in stack and heap space
could also be done manually directly in code. But it seems that people
in the embedded space seem to prefer to write messy linker scripts
instead. Hence this feature proposal:

 --section-order=".... +<NUMBER> "

 is treated as a relative increment of the address.